Added docs and opensbi uboot combined
I followed the bianbu's documentation to generate single itb and then use xyz modem to load it into the ram but the execution stucks after the opensbi is executed. I think FSBL cannot handle the relocation of the itb. So when the u-boot node specifies the execution address to be 0x200000 and the itb is loaded at 0x0, FSBL doesnt relocate the u-boot according to the address specified in the u-boot node